requirements.txt
requests
beautifulsoup4
openpyxl
pandas
ddgs
lxml
python-dotenv
nameparser
langchain
selenium
webdriver-manager
